requirements.txt
django
djangorestframework
django-permission
django-cors-headers
git+git://github.com/sunscrapers/djoser@c54d9f5998a9b542d31b74594acc863cbc01b9ae
bleach
markdown2
django-admin-bootstrapped
django-notifications-hq
django-solo